UI Improvements
Update the top Navbar so that the Search bar is larger and shrinks
appropriately to device size.
Add a Menu "Hamburger" Icon which contains the Import and Help
functions.
Reorder elements in the bottom navbar to be better spaced and
positioned.
Update to the latest Twitter Bootstrap v3.1.1.
Remove some commented out code and reformat.