Multiple enhancements
Add missing features to WhisperNote - menu icon and bottom navbar
changes.
Increase bottom padding to fix bug where bottom was being cutoff in
certain situations.
Add configurable, persistent options to apps to which allow user to
specify a Save Filename and timeout.
Cause apps to timeout if unattended for a period of time.
Fixes #20, #19 and #4.